requirements.txt
Django==1.11.7
olefile==0.44
Pillow==4.3.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
